Moved Position and Command
